Generating a per-resource Service trait and Live class
