+++ zio-k8s-client/src/main/scala/com/coralogix/zio/k8s/client/impl/SubresourceClient.scala
@@ -1,21 +1,18 @@
-package com.coralogix.zio.k8s.client
+package com.coralogix.zio.k8s.client.impl
 
-import com.coralogix.zio.k8s.client.model.{ K8sCluster, K8sNamespace, K8sResourceType }
 import _root_.io.circe._
-import sttp.capabilities.zio.ZioStreams
-import sttp.client3._
+import com.coralogix.zio.k8s.client.model.{ K8sCluster, K8sNamespace, K8sResourceType }
+import com.coralogix.zio.k8s.client.{ K8sFailure, ResourceClientBase, Subresource }
 import sttp.client3.circe._
 import sttp.client3.httpclient.zio._
-import sttp.model.{ StatusCode, Uri }
-
 import zio.IO
 
-class SubresourceClient[T: Encoder: Decoder](
+final class SubresourceClient[T: Encoder: Decoder](
   override protected val resourceType: K8sResourceType,
   override protected val cluster: K8sCluster,
   override protected val backend: SttpClient.Service,
   subresourceName: String
-) extends ResourceClientBase {
+) extends Subresource[T] with ResourceClientBase {
 
   def get(name: String, namespace: Option[K8sNamespace]): IO[K8sFailure, T] =
     handleFailures {
@@ -35,7 +32,7 @@ class SubresourceClient[T: Encoder: Decoder](
       k8sRequest
         .put(modifying(name, Some(subresourceName), namespace, dryRun))
         .body(updatedValue)
-        .response(asJson[T]) // TODO: verify
+        .response(asJson[T])
         .send(backend)
     }
 
@@ -44,7 +41,7 @@ class SubresourceClient[T: Encoder: Decoder](
       k8sRequest
         .post(creating(Some(subresourceName), namespace, dryRun))
         .body(value)
-        .response(asJson[T]) // TODO: verify
+        .response(asJson[T])
         .send(backend)
     }
